move apply_committed to store.
